requirements.txt
dj-database-url==0.3.0
Django==1.9.4
django-webpack-loader==0.3.0
django-countries==3.4.1
whitenoise==2.0.6
gunicorn==19.4.5
django-compressor==2.0
psycopg2==2.6.1
slacker==0.9.9
raven==5.12.0
djangorestframework==3.4.0
Markdown==2.6.6   # Markdown support for the browsable API. todo: migrate to dev only requirements
celery==3.1.23
librabbitmq==1.6.1
scrapinghub==1.7.0
sqlalchemy
plivo==0.11.0
FeinCMS==1.11.4
django-storages==1.5.0
django-taggit==0.18.2
apache-libcloud==0.20.1
boto==2.42.0
redis==2.10.5
simplejson==3.8.2
requests==2.10.0
-e git://github.com/lihan/django-cors-headers.git#egg=django-cors-headers
django-anymail==0.3.1
premailer==3.0.0
pyPdf==1.13
numpy==1.11.0
Wand==0.4.3
Pillow==3.3.0
django-timezone-field==1.3
stripe==1.37.0
django-preventconcurrentlogins==0.7
factory-boy==2.7.0      # todo: dev only dependency
